third_party/WebKit/LayoutTests/webexposed/event-target-in-prototype-expected.txt

Check that interfaces inherit from EventTarget

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS EventSource.prototype.__proto__ is EventTarget.prototype
PASS MessagePort.prototype.__proto__ is EventTarget.prototype
FAIL Node.prototype.__proto__ should be [object Object]. Was [object Object].
PASS WebSocket.prototype.__proto__ is EventTarget.prototype
FAIL Window.prototype.__proto__ should be [object Object]. Was [object Object].
PASS XMLHttpRequest.prototype.__proto__ is EventTarget.prototype
PASS XMLHttpRequestUpload.prototype.__proto__ is EventTarget.prototype
PASS webkitAudioContext.prototype.__proto__ is EventTarget.prototype
PASS successfullyParsed is true

TEST COMPLETE
